+++ io-hotmoka-node-disk/src/main/java/io/hotmoka/node/disk/internal/Mempool.java
@@ -120,8 +120,8 @@ class Mempool {
 	 * The body of the thread that executes requests. Its pops a request from the checked mempool and executes it.
 	 */
 	private void deliver() {
-		int counter = 0;
-		int transactionsPerBlock = node.getConfig().getTransactionsPerBlock();
+		long counter = 0;
+		long transactionsPerBlock = node.getConfig().getTransactionsPerBlock();
 
 		while (!Thread.currentThread().isInterrupted()) {
 			try {
